pyproject.toml
[project]
name = "gx-train"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "augraphy>=8.2.6",
    "fire>=0.7.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "multidict>=6.0.0,!=6.5.0",
    "notebook>=7.4.3",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "ray[tune]>=2.47.1",
    "scikit-learn>=1.7.0",
    "timm>=1.0.15",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "wandb>=0.20.1",
]
